+++ src/ui/draw.rs
@@ -226,8 +226,7 @@ fn prepare_receipts(app: &mut App, height: usize) {
         messages
             .rendered
             .offset
-            .min(selected)
-            .max(selected.saturating_sub(height))
+            .clamp(selected.saturating_sub(height), selected)
     } else {
         messages.rendered.offset
     };
@@ -285,8 +284,7 @@ fn draw_messages<B: Backend>(f: &mut Frame<B>, app: &mut App, area: Rect) {
         messages
             .rendered
             .offset
-            .min(selected)
-            .max(selected.saturating_sub(height))
+            .clamp(selected.saturating_sub(height), selected)
     } else {
         messages.rendered.offset
     };
@@ -317,7 +315,7 @@ fn draw_messages<B: Backend>(f: &mut Frame<B>, app: &mut App, area: Rect) {
                 .expect("non-existent message");
             let date_division = display_date_line(msg.arrived_at, &mut previous_msg_day, width);
             let show_receipt = ShowReceipt::from_msg(&msg, app.user_id, app.config.show_receipts);
-            let msg = display_message(&names, &msg, &prefix, width as usize, height, show_receipt);
+            let msg = display_message(&names, &msg, &prefix, width, height, show_receipt);
 
             [date_division, msg]
         })
@@ -362,7 +360,7 @@ fn draw_messages<B: Backend>(f: &mut Frame<B>, app: &mut App, area: Rect) {
     }
 
     let title: String = if let Some(writing_people) = writing_people {
-        format!("Messages {}", writing_people)
+        format!("Messages {writing_people}")
     } else {
         "Messages".to_string()
     };
@@ -491,7 +489,7 @@ fn display_message(
         .as_ref()
         .and_then(|quote| displayed_quote(names, quote));
     if let Some(quote_text) = quote_text.as_ref() {
-        let quote_prefix = format!("{}> ", prefix);
+        let quote_prefix = format!("{prefix}> ");
         let quote_wrap_opts = textwrap::Options::new(width.saturating_sub(2))
             .initial_indent(&quote_prefix)
             .subsequent_indent(&quote_prefix);
@@ -540,7 +538,7 @@ fn display_message(
     if spans.len() > height {
         // span is too big to be shown fully
         spans.resize(height - 1, Spans::from(""));
-        spans.push(Spans::from(format!("{}[...]", prefix)));
+        spans.push(Spans::from(format!("{prefix}[...]")));
     }
     Some(ListItem::new(Text::from(spans)))
 }